+++ RELEASES.md
@@ -1,4 +1,4 @@
-# Unreleased
+# Version 0.4.0
 
 ## Breaking changes